backend/requirements.txt
# API Server
fastapi
uvicorn[standard]
python-multipart
tortoise-orm

# For HW3 YOLO
ultralytics 
supervision
opencv-python
chardet

# Dev Packages
black
flake8
isort